refactor(cast): inline subscribe implementation
