+++ crates/cast/src/lib.rs
@@ -99,81 +99,6 @@ impl<P: Provider<N> + Clone + Unpin, N: Network> Cast<P, N> {
     pub const fn new(provider: P) -> Self {
         Self { provider, _phantom: PhantomData }
     }
-
-    /// Sets up a subscription to the given filter and writes the logs to the given output.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use alloy_primitives::Address;
-    /// use alloy_provider::{ProviderBuilder, RootProvider, network::AnyNetwork};
-    /// use alloy_rpc_types::Filter;
-    /// use alloy_transport::BoxTransport;
-    /// use cast::Cast;
-    /// use std::{io, str::FromStr};
-    ///
-    /// # async fn foo() -> eyre::Result<()> {
-    /// let provider =
-    ///     ProviderBuilder::<_, _, AnyNetwork>::default().connect("wss://localhost:8545").await?;
-    /// let cast = Cast::new(provider);
-    ///
-    /// let filter =
-    ///     Filter::new().address(Address::from_str("0x00000000006c3852cbEf3e08E8dF289169EdE581")?);
-    /// let mut output = io::stdout();
-    /// cast.subscribe(filter, &mut output).await?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub async fn subscribe(&self, filter: Filter, output: &mut dyn io::Write) -> Result<()> {
-        let mut subscription = self.provider.subscribe_logs(&filter).await?.into_stream();
-
-        // Subscribe to blocks when a `to_block` is set so the stream ends once it is passed.
-        let to_block_number = filter.get_to_block();
-        let mut block_subscription = match to_block_number {
-            Some(_) => Some(self.provider.subscribe_blocks().await?.into_stream()),
-            None => None,
-        };
-
-        let format_json = shell::is_json();
-        if format_json {
-            write!(output, "[")?;
-        }
-
-        let mut first = true;
-        loop {
-            tokio::select! {
-                block = match &mut block_subscription {
-                    Some(bs) => Either::Left(bs.next().fuse()),
-                    None => Either::Right(futures::future::pending()),
-                } => {
-                    if let (Some(block), Some(to_block)) = (block, to_block_number)
-                        && block.number() > to_block
-                    {
-                        break;
-                    }
-                },
-                log = subscription.next() => {
-                    if format_json {
-                        if !first {
-                            write!(output, ",")?;
-                        }
-                        first = false;
-                        write!(output, "{}", serde_json::to_string(&log).unwrap())?;
-                    } else {
-                        writeln!(output, "{}", pretty_log(&log))?;
-                    }
-                },
-                // Break on the cancel signal so the JSON array is still closed.
-                _ = ctrl_c() => break,
-                else => break,
-            }
-        }
-
-        if format_json {
-            write!(output, "]")?;
-        }
-        Ok(())
-    }
 }
 
 /// Renders a log as an indented list item.
